fix: 게시, 수정, 삭제 후 /home으로 이동
